PMIS-AI-Engine/requirements.txt
# PMIS-AI Smart Allocation Engine - Dependencies
# =============================================

# Core Web Framework
flask>=3.0.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & AI
scikit-learn>=1.3.0
sentence-transformers>=3.2.0
torch>=2.2.0
transformers>=4.40.0

# Natural Language Processing
nltk>=3.9.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Additional Dependencies
typing-extensions>=4.13.0
tqdm>=4.67.0
requests>=2.32.0
huggingface-hub>=0.34.0
python-dotenv>=1.0.0

# Development Tools (Optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
